Add file logger RN example
