+++ libs/sdk-react-native/example/package.json
@@ -13,13 +13,18 @@
     "rebuild": "rm -rf node_modules && yarn && yarn pods"
   },
   "dependencies": {
-    "@breeztech/react-native-breez-sdk": "0.5.1-rc6",
+    "@breeztech/react-native-breez-sdk": "0.5.2",
     "@dreson4/react-native-quick-bip39": "^0.0.5",
+    "@react-navigation/native": "6.1.1",
+    "@react-navigation/native-stack": "6.9.7",
     "react": "18.1.0",
     "react-native": "0.70.15",
     "react-native-build-config": "^0.3.2",
+    "react-native-file-logger": "0.4.1",
     "react-native-quick-base64": "^2.0.5",
     "react-native-quick-crypto": "^0.5.0",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "3.7.2",
     "react-native-secure-storage": "https://github.com/satimoto/react-native-secure-storage"
   },
   "devDependencies": {
