+++ libs/sdk-react-native/example/src/components/DebugLine.js
@@ -1,0 +1,15 @@
+import React from "react"
+import { Text, TouchableOpacity, View } from "react-native"
+
+const DebugLine = ({ title, text }) => {
+    return (
+        <TouchableOpacity style={{ flex: 1 }}>
+            <View style={{ margin: 5 }}>
+                <Text style={{ fontWeight: "bold" }}>{title}</Text>
+                {text && text.length > 0 ? <Text>{text}</Text> : <></>}
+            </View>
+        </TouchableOpacity>
+    )
+}
+
+export default DebugLine
