+++ libs/sdk-react-native/example/src/utils/security.ts
