+++ libs/sdk-react-native/example/src/utils/storage.ts
